reorganize marker tests
